+++ test/Unit/FieldDefinition/ReferencesTest.php
@@ -37,7 +37,7 @@ final class ReferencesTest extends AbstractTestCase
      *
      * @param int $count
      */
-    public function testOptionalRejectsInvalidCount(int $count): void
+    public function testConstructorRejectsInvalidCount(int $count): void
     {
         $this->expectException(Exception\InvalidCount::class);
         $this->expectExceptionMessage(\sprintf(
@@ -45,7 +45,7 @@ final class ReferencesTest extends AbstractTestCase
             $count
         ));
 
-        References::optional(
+        new References(
             Fixture\FixtureFactory\Entity\User::class,
             $count
         );
@@ -56,7 +56,7 @@ final class ReferencesTest extends AbstractTestCase
      *
      * @param int $count
      */
-    public function testOptionalResolvesToArrayOfObjectsCreatedByFixtureFactory(int $count): void
+    public function testResolvesToArrayOfObjectsCreatedByFixtureFactory(int $count): void
     {
         $className = Fixture\FixtureFactory\Entity\User::class;
 
@@ -67,61 +67,10 @@ final class ReferencesTest extends AbstractTestCase
 
         $fixtureFactory->define($className);
 
-        $fieldDefinition = References::optional(
+        $fieldDefinition = new References(
             $className,
             $count
         );
-
-        self::assertFalse($fieldDefinition->isRequired());
-
-        $resolved = $fieldDefinition->resolve($fixtureFactory);
-
-        self::assertIsArray($resolved);
-        self::assertCount($count, $resolved);
-        self::assertContainsOnly($className, $resolved);
-    }
-
-    /**
-     * @dataProvider \Ergebnis\FactoryBot\Test\DataProvider\NumberProvider::intLessThanOne()
-     *
-     * @param int $count
-     */
-    public function testRequiredRejectsInvalidCount(int $count): void
-    {
-        $this->expectException(Exception\InvalidCount::class);
-        $this->expectExceptionMessage(\sprintf(
-            'Count needs to be greater than or equal to 1, but %d is not.',
-            $count
-        ));
-
-        References::required(
-            Fixture\FixtureFactory\Entity\User::class,
-            $count
-        );
-    }
-
-    /**
-     * @dataProvider \Ergebnis\FactoryBot\Test\DataProvider\NumberProvider::intGreaterThanOne()
-     *
-     * @param int $count
-     */
-    public function testRequiredResolvesToArrayOfObjectsCreatedByFixtureFactory(int $count): void
-    {
-        $className = Fixture\FixtureFactory\Entity\User::class;
-
-        $fixtureFactory = new FixtureFactory(
-            self::entityManager(),
-            self::faker()
-        );
-
-        $fixtureFactory->define($className);
-
-        $fieldDefinition = References::required(
-            $className,
-            $count
-        );
-
-        self::assertTrue($fieldDefinition->isRequired());
 
         $resolved = $fieldDefinition->resolve($fixtureFactory);